+++ src/commands/newQuartoReprex.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { QUARTO_WIZARD_LOG } from "../constants";
+import { QW_LOG } from "../constants";
 import { showLogsCommand } from "../utils/log";
 import { newQuartoReprex } from "../utils/reprex";
 
@@ -13,7 +13,7 @@ export async function newQuartoReprexCommand(context: vscode.ExtensionContext) {
 		newQuartoReprex(selectedLanguage, context);
 	} else {
 		const message = `No computing language selected. Aborting.`;
-		QUARTO_WIZARD_LOG.appendLine(message);
+		QW_LOG.appendLine(message);
 		vscode.window.showErrorMessage(`${message} ${showLogsCommand()}.`);
 	}
 }